backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Aashutosh Pudasaini <nepalidude3@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"

[tool.poetry.group.dev.dependencies]
aad-fastapi = "^1.1.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
